+++ programs/encore/src/state/mod.rs
@@ -1,11 +1,9 @@
 //! State account definitions
 
 pub mod event_config;
+pub mod nullifier;
 pub mod ticket;
-pub mod identity_counter;
-
 
 pub use event_config::*;
+pub use nullifier::*;
 pub use ticket::*;
-pub use identity_counter::*;
-